Update for recent FFMPEG (removed `-map_channel` option).
Removal of `-map_channel` option was a pain to solve, as now channel
splitting into streams has to happen into a complex filter. For some
reasons, this also did not work well with the 'basic' cut options of
ffmpeg, so had to also move that into the complex filter.

Bright side: changed process to split a whole input audio file in a
single ffmpeg invocation, so process is now several times faster than
before.

Changes:
* Use temp text file to store the complex filter (reduces size of the
  commandline for the ffmpeg invocation).
* Use `channelsplit` to split stereo (or more) multi-channels audio,
  each into their own stream.
* Use `asegment` to generate 5 seconds segments of audio from these
  channel-split streams, each into their own stream.
* Write all these single-channel-5-seconds streams into their own
  destination wav files.